Timer and ammo
Ammo is finished. The player only has 1 bullet and must kill enemies in order to get said ammo back. Timer has been implemented with a 120 second limit. For now, the limit doesn't do anything but I will change that very soon to make it so that the player loses the game if it is reached. Player now respawns if they fall off map.